fix(seeding): widen SMEM read positions from int16_t to int32_t (#142)
Read positions in the SMEM seeding path were stored as int16_t:
query_pos_array (getSMEMsAllPosOneThread / getSMEMsOnePosOneThread /
the lockstep variant, plus the caller's mmc->query_pos_ar),
BatchSlot.start_pos and next_x, and the walk index `x` in
bwtSeedStrategyAllPosOneThread.

For any read longer than INT16_MAX (32767 bp) the position incremented
past 32767 and wrapped negative, so the re-seed loops
(`while (x < readlength)` and the getSMEMsAllPosOneThread do/while)
never advanced past the wrap point. They emitted SMEMs without bound
until matchArray overran its allocation, corrupting the heap and
crashing with SIGSEGV. Short-read workloads never hit this because
their reads stay well under 32767 bp; it only surfaced on long-read
input (ONT/HiFi reads > 32 kb).

Widen all such positions to int32_t. Read length is already an int, so
this is the correct type. Behaviour is unchanged for reads <= 32767 bp:
SAM output is byte-identical to the pre-fix build on Illumina, HiFi and
SBX inputs.

Add a >32767 bp regression case to the long-read suite in
run_unit_tests.sh (a 40 kb read sliced from the checked-in synthetic_1mb
reference; phiX is too small to host the alignment), and widen the
query-position types in the smem_lockstep_parity and smem2 tests to
match the new signatures.

+++ src/bwamem.h
@@ -213,7 +213,7 @@ typedef struct
     int32_t *min_intv_ar[MAX_THREADS];
     int32_t *rid[MAX_THREADS];
     int32_t *lim[MAX_THREADS];
-    int16_t *query_pos_ar[MAX_THREADS];
+    int32_t *query_pos_ar[MAX_THREADS];
     uint8_t *enc_qdb[MAX_THREADS];
     
     int64_t wsize_mem[MAX_THREADS];
